+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.93.0
+* Allow PayPal payment ID and payer ID to be passed during transaction create
+* Add `travel_flight` support to industry-specific data
+* Add `fraud_service_provider` field to `risk_data`
+
 # 2.92.0
 * Add `processor_response_type` to `Transaction`, `AuthorizationAdjustment`, and `CreditCardVerification`.
 * Add `authorization_expires_at` to `Transaction`